RandomBOSS fixes/tweaks, replacement of BOSS with RandomBOSS.
